fix(roads): check if current company owned provided road before execute command handler

+++ road/commands/delete_road_by_oid.py
@@ -1,6 +1,8 @@
 from dataclasses import dataclass
 from uuid import UUID
 
+from rest_framework import exceptions
+
 from core.base.commands import BaseCommand, BaseCommandHandler
 from road.models import Road
 
@@ -8,6 +10,7 @@ from road.models import Road
 @dataclass(frozen=True)
 class DeleteRoadByOidCommand(BaseCommand):
     road_oid: UUID
+    company_name: str
 
 
 @dataclass(eq=False, frozen=True)
@@ -16,5 +19,11 @@ class DeleteRoadByOidCommandHandler(
 ):
     @staticmethod
     def handle(command: DeleteRoadByOidCommand) -> None:
-        road = Road.objects.get(oid=command.road_oid)
+        company_roads = Road.objects.filter(company__name=command.company_name)
+        road = company_roads.filter(oid=command.road_oid).first()
+        if not road:
+            raise exceptions.NotFound(
+                detail=f"road with {command.road_oid} oid not found",
+            )
+
         road.delete()
